+++ packages/web-components/core/src/animation/index.ts
@@ -1,2 +1,3 @@
 export * from "./controlsToPromise.js";
 export * from "./awaitAllAnimations.js";
+export * from "./getTransitionConfig.js";