+++ packages/web-components/parallax/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/web-component-parallax
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-core@1.0.2
+
 ## 1.0.3
 
 ### Patch Changes